+++ packages/relax/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aiszlab/relax
 
+## 1.3.4
+
+### Patch Changes
+
+- - `wasp`.`dev` command add url print outputs
+  - `bee` bootstrap api improve, use `router`.`createBrowserRouter`
+  - `bee` add `router` & `storage` exports
+
 ## 1.3.3
 
 ### Patch Changes